refactor: ♻️ code cleanup
