+++ src/components/availability/group-availability.tsx
@@ -1,12 +1,14 @@
 "use client";
 
 import { alpha, useTheme } from "@mui/material/styles";
-import React, { useMemo } from "react";
+import { Fragment, useMemo } from "react";
 import { useShallow } from "zustand/shallow";
 import { GroupAvailabilityBlock } from "@/components/availability/group-availability-block";
 import type { GridCellHandlers } from "@/components/availability/table/availability-block-cell";
+import { AvailabilityTimeTicks } from "@/components/availability/table/availability-time-ticks";
 import {
 	formatScheduledTimeRange,
+	generateCellKey,
 	generateDateKey,
 	getTimestampFromBlockIndex,
 	spacerBeforeDate,
@@ -24,6 +26,7 @@ function calculateBlockColor({
 	showBestTimes,
 	maxAvailability,
 	primaryColor,
+	ifNeededColor,
 	ifNeededBlock,
 }: {
 	block: string[];
@@ -33,6 +36,7 @@ function calculateBlockColor({
 	showBestTimes: boolean;
 	maxAvailability: number;
 	primaryColor: string;
+	ifNeededColor: string;
 	ifNeededBlock: string[];
 }): string {
 	if (selectedMembers.length) {
@@ -45,28 +49,28 @@ function calculateBlockColor({
 
 		if (selectedInBlock.length) {
 			const proportion = selectedInBlock.length / selectedMembers.length;
-			return alpha(primaryColor, proportion); // available = pink
+			return alpha(primaryColor, proportion);
 		}
 		if (ifNeededInBlock.length) {
 			const proportion = ifNeededInBlock.length / selectedMembers.length;
-			return `hsl(var(--if-needed) / ${proportion})`;
+			return alpha(ifNeededColor, proportion);
 		}
 		return "transparent";
 	}
 
 	if (hoveredMember) {
 		if (block.includes(hoveredMember)) {
-			return alpha(primaryColor, 1); // available = pink
+			return alpha(primaryColor, 1);
 		}
 		if (ifNeededBlock.includes(hoveredMember)) {
-			return "hsl(var(--if-needed))";
+			return ifNeededColor;
 		}
 		return "transparent";
 	}
 
 	if (showBestTimes) {
 		if (block.length === maxAvailability && maxAvailability > 0) {
-			return "hsl(var(--primary))";
+			return primaryColor;
 		}
 		return "transparent";
 	}
@@ -74,11 +78,11 @@ function calculateBlockColor({
 	if (numMembers) {
 		if (ifNeededBlock.length > 0) {
 			const opacity = ifNeededBlock.length / numMembers;
-			return `hsl(var(--if-needed) / ${opacity})`;
+			return alpha(ifNeededColor, opacity);
 		}
 		if (block.length > 0) {
 			const opacity = block.length / numMembers;
-			return alpha(primaryColor, opacity); // available = pink
+			return alpha(primaryColor, opacity);
 		}
 	}
 
@@ -120,8 +124,6 @@ function edgesFor(
 interface GroupAvailabilityProps {
 	meetingId: string;
 	meetingTitle?: string;
-	timeBlock: number;
-	blockIndex: number;
 	availabilityTimeBlocks: number[];
 	fromTime: number;
 	availabilityDates: ZotDate[];
@@ -138,8 +140,6 @@ interface GroupAvailabilityProps {
 
 export function GroupAvailability({
 	meetingTitle,
-	timeBlock,
-	blockIndex,
 	availabilityTimeBlocks,
 	fromTime,
 	availabilityDates,
@@ -151,39 +151,40 @@ export function GroupAvailability({
 	handlers,
 }: GroupAvailabilityProps) {
 	const theme = useTheme();
+	const primaryColor = theme.palette.primary.main;
+	const ifNeededColor = theme.palette.ifNeeded.main;
 
-	const { currentPage, itemsPerPage } = useAvailabilityStore(
+	const {
+		currentPage,
+		itemsPerPage,
+		hoveredMember,
+		selectedMembers,
+		draftRange,
+		hoverRange,
+		committedRange,
+		scheduledTimes,
+		pendingAdds,
+		pendingRemovals,
+		showBestTimes,
+		isScheduled,
+	} = useAvailabilityStore(
 		useShallow((state) => ({
 			currentPage: state.currentPage,
 			itemsPerPage: state.itemsPerPage,
-		})),
-	);
-
-	const { hoveredMember, selectedMembers } = useAvailabilityStore(
-		useShallow((state) => ({
 			hoveredMember: state.hoveredMember,
 			selectedMembers: state.selectedMembers,
-		})),
-	);
-
-	const { draftRange, hoverRange, committedRange } = useAvailabilityStore(
-		useShallow((state) => ({
 			draftRange: state.draftRange,
 			hoverRange: state.hoverRange,
 			committedRange: state.committedRange,
+			scheduledTimes: state.scheduledTimes,
+			pendingAdds: state.pendingAdds,
+			pendingRemovals: state.pendingRemovals,
+			showBestTimes: state.enabled,
+			isScheduled: state.isScheduled,
 		})),
 	);
 
 	const numMembers = members.length;
-	const showBestTimes = useAvailabilityStore((state) => state.enabled);
-
-	const { scheduledTimes, pendingAdds, pendingRemovals } = useAvailabilityStore(
-		useShallow((state) => ({
-			scheduledTimes: state.scheduledTimes,
-			pendingAdds: state.pendingAdds,
-			pendingRemovals: state.pendingRemovals,
-		})),
-	);
 
 	const { scheduledTimeRange, scheduledBlockCount } = useMemo(() => {
 		const effective = new Set([...scheduledTimes, ...pendingAdds]);
@@ -198,7 +199,6 @@ export function GroupAvailability({
 
 	const maxAvailability = useMemo(() => {
 		if (!showBestTimes || numMembers === 0) return 0;
-
 		let max = 0;
 		availabilityDates.forEach((date) => {
 			Object.values(date.groupAvailability).forEach((memberIds) => {
@@ -208,154 +208,183 @@ export function GroupAvailability({
 		return max;
 	}, [showBestTimes, numMembers, availabilityDates]);
 
-	const isScheduled = useAvailabilityStore((state) => state.isScheduled);
-
-	const isTopOfHour = timeBlock % 60 === 0;
-	const isHalfHour = timeBlock % 60 === 30;
-	const isLastRow = blockIndex === availabilityTimeBlocks.length - 1;
-
-	const spacers = spacerBeforeDate(currentPageAvailability["availabilities"]);
-
-	return currentPageAvailability.availabilities.map(
-		(selectedDate, pageDateIndex) => {
-			const ifNeededDate = currentPageAvailability.ifNeeded[pageDateIndex];
-
-			const key = generateDateKey({
-				selectedDate,
-				timeBlock,
-				pageDateIndex,
-			});
-
-			if (selectedDate) {
-				const zotDateIndex = pageDateIndex + currentPage * itemsPerPage;
-
-				const draftEdges = edgesFor(
-					draftRange,
-					zotDateIndex,
-					blockIndex,
-					"draft",
-				);
-				const hoverEdges =
-					draftEdges === null
-						? edgesFor(hoverRange, zotDateIndex, blockIndex, "hover")
-						: null;
-				const committedEdges =
-					draftEdges === null && hoverEdges === null && !isScheduling
-						? edgesFor(committedRange, zotDateIndex, blockIndex, "committed")
-						: null;
-				const selectionEdges =
-					draftEdges ?? hoverEdges ?? committedEdges ?? null;
-
-				const timestamp = getTimestampFromBlockIndex(
-					blockIndex,
-					zotDateIndex,
-					fromTime,
-					availabilityDates,
-					timeZone,
-				);
-
-				const block = selectedDate.groupAvailability[timestamp] || [];
-				const ifNeededBlock = ifNeededDate?.groupAvailability[timestamp] || [];
-				const blockColor = calculateBlockColor({
-					block,
-					hoveredMember,
-					selectedMembers,
-					numMembers,
-					showBestTimes,
-					maxAvailability,
-					ifNeededBlock,
-					primaryColor: theme.palette.primary.main,
-				});
-				const blockIsScheduled = isScheduled(timestamp);
-
-				const prevTimestamp =
-					blockIndex > 0
-						? getTimestampFromBlockIndex(
-								blockIndex - 1,
-								zotDateIndex,
-								fromTime,
-								availabilityDates,
-								timeZone,
-							)
-						: "";
-				const nextTimestamp =
-					blockIndex < availabilityTimeBlocks.length - 1
-						? getTimestampFromBlockIndex(
-								blockIndex + 1,
-								zotDateIndex,
-								fromTime,
-								availabilityDates,
-								timeZone,
-							)
-						: "";
-				const isTopEdge =
-					blockIsScheduled && (!prevTimestamp || !isScheduled(prevTimestamp));
-				const isBottomEdge =
-					blockIsScheduled && (!nextTimestamp || !isScheduled(nextTimestamp));
-
-				const tableCellStyles = cn(
-					isTopOfHour ? "border-t-[1px] border-t-gray-base" : "",
-					isHalfHour
-						? "border-t border-t-gray-base [border-top-style:dotted]"
-						: "",
-					isLastRow ? "border-b-[1px]" : "",
-				);
-
-				return (
-					<React.Fragment key={key}>
-						{spacers[pageDateIndex] && (
-							<td className="w-3 md:w-4" aria-hidden="true" />
-						)}
-						<td className={cn("px-0 py-0", isTopEdge && "relative z-[1]")}>
-							<GroupAvailabilityBlock
-								className={cn(
-									"group-availability-block block",
-									isScheduling && "cursor-row-resize [touch-action:pinch-zoom]",
-								)}
-								onPointerDown={handlers.onPointerDown}
-								onPointerMove={handlers.onPointerMove}
-								onPointerUp={handlers.onPointerUp}
-								onPointerCancel={handlers.onPointerCancel}
-								onKeyDown={handlers.onKeyDown}
-								onHover={() =>
-									handlers.onCellHover?.({
-										zotDateIndex,
-										blockIndex,
-									})
-								}
-								blockColor={blockColor}
-								isScheduled={blockIsScheduled}
-								isScheduledTopEdge={isTopEdge}
-								isScheduledBottomEdge={isBottomEdge}
-								scheduledMeetingTitle={isTopEdge ? meetingTitle : undefined}
-								scheduledTimeRange={isTopEdge ? scheduledTimeRange : undefined}
-								scheduledBlockCount={
-									isTopEdge ? scheduledBlockCount : undefined
-								}
-								tableCellStyles={tableCellStyles}
-								hasSpacerBefore={spacers[pageDateIndex]}
-								dateIndex={zotDateIndex}
-								blockIndex={blockIndex}
-								selectionEdges={selectionEdges}
-							/>
-						</td>
-					</React.Fragment>
-				);
-			} else {
-				return (
-					// Because these elements are hidden spacers, we consider mouse hovers to be "leaving" the table
-					<React.Fragment key={key}>
-						{spacers[pageDateIndex] && (
-							<td
-								className="w-3 md:w-4"
-								aria-hidden="true"
-								onMouseEnter={onMouseLeave}
-							/>
-						)}
-						<td onMouseEnter={onMouseLeave}></td>
-					</React.Fragment>
+	const timestampsByCell = useMemo(() => {
+		const map = new Map<string, string>();
+		for (let d = 0; d < availabilityDates.length; d++) {
+			for (let b = 0; b < availabilityTimeBlocks.length; b++) {
+				map.set(
+					generateCellKey(d, b),
+					getTimestampFromBlockIndex(
+						b,
+						d,
+						fromTime,
+						availabilityDates,
+						timeZone,
+					),
 				);
 			}
-		},
+		}
+		return map;
+	}, [availabilityDates, availabilityTimeBlocks, fromTime, timeZone]);
+
+	const spacers = spacerBeforeDate(currentPageAvailability.availabilities);
+	const lastRowIndex = availabilityTimeBlocks.length - 1;
+
+	return (
+		<>
+			{availabilityTimeBlocks.map((timeBlock, blockIndex) => {
+				const isTopOfHour = timeBlock % 60 === 0;
+				const isHalfHour = timeBlock % 60 === 30;
+				const isLastRow = blockIndex === lastRowIndex;
+
+				return (
+					<tr key={`block-${timeBlock}`}>
+						<AvailabilityTimeTicks timeBlock={timeBlock} />
+						{currentPageAvailability.availabilities.map(
+							(selectedDate, pageDateIndex) => {
+								const ifNeededDate =
+									currentPageAvailability.ifNeeded[pageDateIndex];
+								const key = generateDateKey({
+									selectedDate,
+									timeBlock,
+									pageDateIndex,
+								});
+
+								if (!selectedDate) {
+									// Hidden spacer columns treat hover-in as a grid-leave.
+									return (
+										<Fragment key={key}>
+											{spacers[pageDateIndex] && (
+												<td
+													className="w-3 md:w-4"
+													aria-hidden="true"
+													onMouseEnter={onMouseLeave}
+												/>
+											)}
+											<td onMouseEnter={onMouseLeave}></td>
+										</Fragment>
+									);
+								}
+
+								const zotDateIndex = pageDateIndex + currentPage * itemsPerPage;
+
+								const draftEdges = edgesFor(
+									draftRange,
+									zotDateIndex,
+									blockIndex,
+									"draft",
+								);
+								const hoverEdges =
+									draftEdges === null
+										? edgesFor(hoverRange, zotDateIndex, blockIndex, "hover")
+										: null;
+								const committedEdges =
+									draftEdges === null && hoverEdges === null && !isScheduling
+										? edgesFor(
+												committedRange,
+												zotDateIndex,
+												blockIndex,
+												"committed",
+											)
+										: null;
+								const selectionEdges =
+									draftEdges ?? hoverEdges ?? committedEdges ?? null;
+
+								const timestamp =
+									timestampsByCell.get(
+										generateCellKey(zotDateIndex, blockIndex),
+									) ?? "";
+
+								const block = selectedDate.groupAvailability[timestamp] || [];
+								const ifNeededBlock =
+									ifNeededDate?.groupAvailability[timestamp] || [];
+								const blockColor = calculateBlockColor({
+									block,
+									hoveredMember,
+									selectedMembers,
+									numMembers,
+									showBestTimes,
+									maxAvailability,
+									ifNeededBlock,
+									primaryColor,
+									ifNeededColor,
+								});
+								const blockIsScheduled = isScheduled(timestamp);
+
+								const prevTimestamp =
+									blockIndex > 0
+										? (timestampsByCell.get(
+												generateCellKey(zotDateIndex, blockIndex - 1),
+											) ?? "")
+										: "";
+								const nextTimestamp =
+									blockIndex < availabilityTimeBlocks.length - 1
+										? (timestampsByCell.get(
+												generateCellKey(zotDateIndex, blockIndex + 1),
+											) ?? "")
+										: "";
+								const isTopEdge =
+									blockIsScheduled &&
+									(!prevTimestamp || !isScheduled(prevTimestamp));
+								const isBottomEdge =
+									blockIsScheduled &&
+									(!nextTimestamp || !isScheduled(nextTimestamp));
+
+								const tableCellStyles = cn(
+									isTopOfHour ? "border-t-[1px] border-t-gray-base" : "",
+									isHalfHour
+										? "border-t border-t-gray-base [border-top-style:dotted]"
+										: "",
+									isLastRow ? "border-b-[1px]" : "",
+								);
+
+								return (
+									<Fragment key={key}>
+										{spacers[pageDateIndex] && (
+											<td className="w-3 md:w-4" aria-hidden="true" />
+										)}
+										<td
+											className={cn("px-0 py-0", isTopEdge && "relative z-[1]")}
+										>
+											<GroupAvailabilityBlock
+												className={cn(
+													"group-availability-block block",
+													isScheduling &&
+														"cursor-row-resize [touch-action:pinch-zoom]",
+												)}
+												onPointerDown={handlers.onPointerDown}
+												onPointerMove={handlers.onPointerMove}
+												onPointerUp={handlers.onPointerUp}
+												onPointerCancel={handlers.onPointerCancel}
+												onKeyDown={handlers.onKeyDown}
+												onHoverCell={handlers.onCellHover}
+												blockColor={blockColor}
+												isScheduled={blockIsScheduled}
+												isScheduledTopEdge={isTopEdge}
+												isScheduledBottomEdge={isBottomEdge}
+												scheduledMeetingTitle={
+													isTopEdge ? meetingTitle : undefined
+												}
+												scheduledTimeRange={
+													isTopEdge ? scheduledTimeRange : undefined
+												}
+												scheduledBlockCount={
+													isTopEdge ? scheduledBlockCount : undefined
+												}
+												tableCellStyles={tableCellStyles}
+												hasSpacerBefore={spacers[pageDateIndex]}
+												dateIndex={zotDateIndex}
+												blockIndex={blockIndex}
+												selectionEdges={selectionEdges}
+											/>
+										</td>
+									</Fragment>
+								);
+							},
+						)}
+					</tr>
+				);
+			})}
+		</>
 	);
 }
