+++ src/components/availability/table/availability-block.tsx
@@ -3,46 +3,26 @@ import {
 	type PaintMode,
 	paintWillChange,
 } from "@/lib/availability/paint-selection";
-import type { SelectionStateType } from "@/lib/types/availability";
 import { cn } from "@/lib/utils";
 
 interface AvailabilityBlockProps {
 	isAvailable: boolean;
 	isIfNeeded: boolean;
-	zotDateIndex: number;
-	blockIndex: number;
-	draftRange: SelectionStateType | undefined;
+	isInDraftRange: boolean;
 	paintMode: PaintMode;
 	showImportPreview?: boolean;
-}
-
-function covers(
-	range: SelectionStateType | undefined,
-	zotDateIndex: number,
-	blockIndex: number,
-): boolean {
-	if (!range) return false;
-	return (
-		range.earlierDateIndex <= zotDateIndex &&
-		zotDateIndex <= range.laterDateIndex &&
-		range.earlierBlockIndex <= blockIndex &&
-		blockIndex <= range.laterBlockIndex
-	);
 }
 
 export function AvailabilityBlock({
 	isAvailable,
 	isIfNeeded,
-	zotDateIndex,
-	blockIndex,
-	draftRange,
+	isInDraftRange,
 	paintMode,
 	showImportPreview = false,
 }: AvailabilityBlockProps) {
 	const backgroundColor = useMemo(() => {
 		const showDraftOverlay =
-			covers(draftRange, zotDateIndex, blockIndex) &&
-			paintWillChange(paintMode, { isAvailable, isIfNeeded });
+			isInDraftRange && paintWillChange(paintMode, { isAvailable, isIfNeeded });
 
 		if (showDraftOverlay) return "bg-primary/40";
 		return isAvailable
@@ -50,14 +30,7 @@ export function AvailabilityBlock({
 			: isIfNeeded
 				? "bg-if-needed"
 				: "transparent";
-	}, [
-		draftRange,
-		paintMode,
-		isAvailable,
-		isIfNeeded,
-		zotDateIndex,
-		blockIndex,
-	]);
+	}, [isInDraftRange, paintMode, isAvailable, isIfNeeded]);
 
 	return (
 		<div className="pointer-events-none relative block h-full w-full py-2">
