+++ src/components/availability/table/availability-blocks.tsx
@@ -11,9 +11,29 @@ import {
 	getTimestampFromBlockIndex,
 	spacerBeforeDate,
 } from "@/lib/availability/utils";
-import type { ProcessedCellEventSegments } from "@/lib/types/availability";
+import type {
+	EventSegment,
+	ProcessedCellEventSegments,
+	SelectionStateType,
+} from "@/lib/types/availability";
 import type { ZotDate } from "@/lib/zotdate";
 import { useAvailabilityStore } from "@/store/useAvailabilityStore";
+
+const EMPTY_EVENT_SEGMENTS: EventSegment[] = [];
+
+function coversCell(
+	range: SelectionStateType | undefined,
+	zotDateIndex: number,
+	blockIndex: number,
+): boolean {
+	if (!range) return false;
+	return (
+		range.earlierDateIndex <= zotDateIndex &&
+		zotDateIndex <= range.laterDateIndex &&
+		range.earlierBlockIndex <= blockIndex &&
+		blockIndex <= range.laterBlockIndex
+	);
+}
 
 interface AvailabilityBlocksProps extends GridCellHandlers {
 	timeBlock: number;
@@ -46,10 +66,15 @@ export function AvailabilityBlocks({
 	onPointerCancel,
 	onKeyDown,
 }: AvailabilityBlocksProps) {
-	const { currentPage, itemsPerPage } = useAvailabilityStore(
+	// Single subscription per row. Cells receive a derived boolean, not the
+	// range itself — this is the one place that still needs to react to
+	// draftRange changes, so the re-render surface is bounded to the rows
+	// whose membership flipped.
+	const { currentPage, itemsPerPage, draftRange } = useAvailabilityStore(
 		useShallow((state) => ({
 			currentPage: state.currentPage,
 			itemsPerPage: state.itemsPerPage,
+			draftRange: state.draftRange,
 		})),
 	);
 	const importPreviewIsoSet = useAvailabilityStore(
@@ -60,7 +85,7 @@ export function AvailabilityBlocks({
 	const isHalfHour = timeBlock % 60 === 30;
 	const isLastRow = blockIndex === availabilityTimeBlocksLength - 1;
 
-	const spacers = spacerBeforeDate(currentPageAvailability["availabilities"]);
+	const spacers = spacerBeforeDate(currentPageAvailability.availabilities);
 
 	return (
 		<>
@@ -82,7 +107,8 @@ export function AvailabilityBlocks({
 							(ifNeededDate?.getBlockAvailability(blockIndex) ?? false);
 
 						const cellKey = generateCellKey(zotDateIndex, blockIndex);
-						const segmentsForCell = processedCellSegments.get(cellKey) || [];
+						const segmentsForCell =
+							processedCellSegments.get(cellKey) ?? EMPTY_EVENT_SEGMENTS;
 
 						const slotIso = getTimestampFromBlockIndex(
 							blockIndex,
@@ -93,6 +119,12 @@ export function AvailabilityBlocks({
 						);
 						const showImportPreview =
 							Boolean(slotIso) && Boolean(importPreviewIsoSet?.has(slotIso));
+
+						const isInDraftRange = coversCell(
+							draftRange,
+							zotDateIndex,
+							blockIndex,
+						);
 
 						return (
 							<Fragment key={key}>
@@ -110,6 +142,7 @@ export function AvailabilityBlocks({
 									eventSegments={segmentsForCell}
 									hasSpacerBefore={spacers[pageDateIndex]}
 									showImportPreview={showImportPreview}
+									isInDraftRange={isInDraftRange}
 									paintMode={paintMode}
 									onPointerDown={onPointerDown}
 									onPointerMove={onPointerMove}
